--- aeon/performance_metrics/forecasting/__init__.py
@@ -1,55 +1,0 @@
-"""Metrics to assess performance on forecasting task.
-
-Functions named as ``*_score`` return a scalar value to maximize: the higher
-the better.
-Function named as ``*_error`` or ``*_loss`` return a scalar value to minimize:
-the lower the better.
-"""
-
-__all__ = [
-    "mean_absolute_scaled_error",
-    "median_absolute_scaled_error",
-    "mean_squared_scaled_error",
-    "median_squared_scaled_error",
-    "mean_absolute_error",
-    "mean_squared_error",
-    "median_absolute_error",
-    "median_squared_error",
-    "geometric_mean_absolute_error",
-    "geometric_mean_squared_error",
-    "mean_absolute_percentage_error",
-    "median_absolute_percentage_error",
-    "mean_squared_percentage_error",
-    "median_squared_percentage_error",
-    "mean_relative_absolute_error",
-    "median_relative_absolute_error",
-    "geometric_mean_relative_absolute_error",
-    "geometric_mean_relative_squared_error",
-    "mean_asymmetric_error",
-    "mean_linex_error",
-    "relative_loss",
-]
-
-from aeon.performance_metrics.forecasting._functions import (
-    geometric_mean_absolute_error,
-    geometric_mean_relative_absolute_error,
-    geometric_mean_relative_squared_error,
-    geometric_mean_squared_error,
-    mean_absolute_error,
-    mean_absolute_percentage_error,
-    mean_absolute_scaled_error,
-    mean_asymmetric_error,
-    mean_linex_error,
-    mean_relative_absolute_error,
-    mean_squared_error,
-    mean_squared_percentage_error,
-    mean_squared_scaled_error,
-    median_absolute_error,
-    median_absolute_percentage_error,
-    median_absolute_scaled_error,
-    median_relative_absolute_error,
-    median_squared_error,
-    median_squared_percentage_error,
-    median_squared_scaled_error,
-    relative_loss,
-)
