--- aeon/performance_metrics/forecasting/tests/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for aeon performance metrics module."""
