--- aeon/performance_metrics/forecasting/tests/test_metrics.py
@@ -1,56 +1,0 @@
-"""Tests for some metrics."""
-
-__maintainer__ = []
-
-import numpy as np
-
-from aeon.performance_metrics.forecasting import (
-    geometric_mean_squared_error,
-    mean_linex_error,
-)
-
-
-def test_gmse_function():
-    """Doctest from geometric_mean_squared_error."""
-    gmse = geometric_mean_squared_error
-    y_true = np.array([3, -0.5, 2, 7, 2])
-    y_pred = np.array([2.5, 0.0, 2, 8, 1.25])
-    assert np.allclose(gmse(y_true, y_pred), 2.80399089461488e-07)
-    assert np.allclose(gmse(y_true, y_pred, square_root=True), 0.000529527232030127)
-    y_true = np.array([[0.5, 1], [-1, 1], [7, -6]])
-    y_pred = np.array([[0, 2], [-1, 2], [8, -5]])
-    assert np.allclose(gmse(y_true, y_pred), 0.5000000000115499)
-    assert np.allclose(gmse(y_true, y_pred, square_root=True), 0.5000024031086919)
-    assert np.allclose(
-        gmse(y_true, y_pred, multioutput="raw_values"),
-        np.array([2.30997255e-11, 1.00000000e00]),
-    )
-    assert np.allclose(
-        gmse(y_true, y_pred, multioutput="raw_values", square_root=True),
-        np.array([4.80621738e-06, 1.00000000e00]),
-    )
-    assert np.allclose(gmse(y_true, y_pred, multioutput=[0.3, 0.7]), 0.7000000000069299)
-    assert np.allclose(
-        gmse(y_true, y_pred, multioutput=[0.3, 0.7], square_root=True),
-        0.7000014418652152,
-    )
-
-
-def test_linex_function():
-    """Test from mean_linex_error."""
-    y_true = np.array([3, -0.5, 2, 7, 2])
-    y_pred = np.array([2.5, 0.0, 2, 8, 1.25])
-    assert np.allclose(mean_linex_error(y_true, y_pred), 0.19802627763937575)
-    assert np.allclose(mean_linex_error(y_true, y_pred, b=2), 0.3960525552787515)
-    assert np.allclose(mean_linex_error(y_true, y_pred, a=-1), 0.2391800623225643)
-    y_true = np.array([[0.5, 1], [-1, 1], [7, -6]])
-    y_pred = np.array([[0, 2], [-1, 2], [8, -5]])
-    assert np.allclose(mean_linex_error(y_true, y_pred), 0.2700398392309829)
-    assert np.allclose(mean_linex_error(y_true, y_pred, a=-1), 0.49660966225813563)
-    assert np.allclose(
-        mean_linex_error(y_true, y_pred, multioutput="raw_values"),
-        np.array([0.17220024, 0.36787944]),
-    )
-    assert np.allclose(
-        mean_linex_error(y_true, y_pred, multioutput=[0.3, 0.7]), 0.30917568000716666
-    )
