+++ aeon/performance_metrics/tests/test_forecasting_metrics.py
@@ -1,6 +1,6 @@
 """Tests for forecasting performance metrics."""
 
-__maintainer__ = []
+__maintainer__ = ["TonyBagnall"]
 
 import inspect
 
@@ -32,7 +32,6 @@ from aeon.performance_metrics.forecasting import (
     median_squared_scaled_error,
     relative_loss,
 )
-from aeon.performance_metrics.forecasting._functions import _get_kwarg
 from aeon.testing.data_generation._legacy import make_series
 
 RANDOM_SEED = 42
@@ -501,12 +500,6 @@ def test_y_true_y_pred_inconsistent_n_variables_raises_error(metric_func_name):
         metric_func(y_true, y_pred, y_train=y_train, y_pred_benchmark=y_pred_benchmark)
 
 
-def test_kwargs():
-    """Test get_kwarg with None."""
-    with pytest.raises(ValueError):
-        _get_kwarg(None)
-
-
 functions = [
     median_squared_scaled_error,
     mean_squared_error,
@@ -556,3 +549,49 @@ def test_check_inputs(function):
     )
     with pytest.raises(ValueError):
         function(y_true, y_pred, **kwargs)
+
+
+def test_gmse_function():
+    """Doctest from geometric_mean_squared_error."""
+    gmse = geometric_mean_squared_error
+    y_true = np.array([3, -0.5, 2, 7, 2])
+    y_pred = np.array([2.5, 0.0, 2, 8, 1.25])
+    assert np.allclose(gmse(y_true, y_pred), 2.80399089461488e-07)
+    assert np.allclose(gmse(y_true, y_pred, square_root=True), 0.000529527232030127)
+    y_true = np.array([[0.5, 1], [-1, 1], [7, -6]])
+    y_pred = np.array([[0, 2], [-1, 2], [8, -5]])
+    assert np.allclose(gmse(y_true, y_pred), 0.5000000000115499)
+    assert np.allclose(gmse(y_true, y_pred, square_root=True), 0.5000024031086919)
+    assert np.allclose(
+        gmse(y_true, y_pred, multioutput="raw_values"),
+        np.array([2.30997255e-11, 1.00000000e00]),
+    )
+    assert np.allclose(
+        gmse(y_true, y_pred, multioutput="raw_values", square_root=True),
+        np.array([4.80621738e-06, 1.00000000e00]),
+    )
+    assert np.allclose(gmse(y_true, y_pred, multioutput=[0.3, 0.7]), 0.7000000000069299)
+    assert np.allclose(
+        gmse(y_true, y_pred, multioutput=[0.3, 0.7], square_root=True),
+        0.7000014418652152,
+    )
+
+
+def test_linex_function():
+    """Test from mean_linex_error."""
+    y_true = np.array([3, -0.5, 2, 7, 2])
+    y_pred = np.array([2.5, 0.0, 2, 8, 1.25])
+    assert np.allclose(mean_linex_error(y_true, y_pred), 0.19802627763937575)
+    assert np.allclose(mean_linex_error(y_true, y_pred, b=2), 0.3960525552787515)
+    assert np.allclose(mean_linex_error(y_true, y_pred, a=-1), 0.2391800623225643)
+    y_true = np.array([[0.5, 1], [-1, 1], [7, -6]])
+    y_pred = np.array([[0, 2], [-1, 2], [8, -5]])
+    assert np.allclose(mean_linex_error(y_true, y_pred), 0.2700398392309829)
+    assert np.allclose(mean_linex_error(y_true, y_pred, a=-1), 0.49660966225813563)
+    assert np.allclose(
+        mean_linex_error(y_true, y_pred, multioutput="raw_values"),
+        np.array([0.17220024, 0.36787944]),
+    )
+    assert np.allclose(
+        mean_linex_error(y_true, y_pred, multioutput=[0.3, 0.7]), 0.30917568000716666
+    )
